changed game description and updated game

+++ game/js/main-scene.js
@@ -1,10 +1,18 @@
+import Player from './player.js';
+import Enemy from './enemy.js';
+import { randomRangeInt } from './script.js';
+
 class MainScene extends Phaser.Scene {
 	constructor() {
 		super('MainScene');
 		this.deltaTime, this.cursors, this.ground, this.player, this.scoreText;
-		this.score = 0;
+
+		this.timer = 180;
+		this.score;
+
 		this.enemies = [];
 		this.enemySpawnTimer = 180;
+
 		this.background = {
 			sky: {
 				image: undefined,
@@ -13,7 +21,7 @@ class MainScene extends Phaser.Scene {
 					x: 0,
 					y: 0,
 				},
-				scrollSpeed: 0.6,
+				scrollSpeed: 0.4,
 			},
 			ground: {
 				image: undefined,
@@ -22,12 +30,14 @@ class MainScene extends Phaser.Scene {
 					x: 0,
 					y: 0,
 				},
-				scrollSpeed: 0.75,
+				scrollSpeed: 0.8,
 			},
 		};
 	}
 
 	preload() {
+		this.score = 0;
+
 		this.background.sky.image = this.load.image('sky', 'img/background.png');
 		this.background.ground.image = this.load.image('ground', 'img/ground.png');
 
@@ -40,8 +50,6 @@ class MainScene extends Phaser.Scene {
 	}
 
 	create() {
-		this.physics.world.setFPS(120);
-
 		// Create ground
 		this.platform = this.physics.add.staticGroup();
 		this.platform
@@ -49,29 +57,32 @@ class MainScene extends Phaser.Scene {
 			.setScale(20, 0.2)
 			.refreshBody();
 
-		// Initialize player and background
+		// Draw player and background
 		this.background.sky.tile = this.add.tileSprite(this.background.sky.pos.x, this.background.sky.pos.y, this.background.sky.image.width, this.background.sky.image.height, 'sky').setOrigin(0, 0);
 		if (!this.player.isDead) this.player.create();
 		this.background.ground.tile = this.add.tileSprite(this.background.ground.pos.x, this.background.ground.pos.y, this.background.ground.image.width, this.background.ground.image.height, 'ground').setOrigin(0, 0);
 
-		this.scoreText = this.add.text(/* this.sys.game.canvas.width - */ 20, 5, this.score, {
-			fontFamily: 'VT323',
-			fontSize: 24,
+		// Draw text
+		this.scoreText = this.add.text(20, 5, ':D', {
+			fontFamily: 'pixel font',
+			fontSize: 9,
 		});
 	}
 
 	update(time, delta) {
 		let realTime = delta / 1000; // Get delta in ms
 
-		this.scoreText.text = Math.round(this.score);
+		this.timer += realTime * 1.5;
+
+		this.scoreText.text = 'SCORE: ' + Math.round(this.score);
 
 		if (!this.player.isDead) {
-			this.score += realTime * 20;
+			this.score += realTime * 10;
 
 			// Spawn enemies
-			this.enemySpawnTimer--;
+			this.enemySpawnTimer -= 205 * realTime;
 			if (this.enemySpawnTimer < 0) {
-				this.enemySpawnTimer = randomRangeInt(60, 180);
+				this.enemySpawnTimer = randomRangeInt(100, 200);
 				this.enemies.push(new Enemy(this.sys.game.canvas.width + 8, 136, this));
 			}
 
@@ -88,13 +99,21 @@ class MainScene extends Phaser.Scene {
 
 				// Kill player
 				let that = this;
-				this.physics.add.overlap(this.player.collider, this.enemies[i].collider, function () {
-					that.player.kill();
-				});
+				if (this.enemies.length > 0) {
+					this.physics.add.overlap(this.player.collider, this.enemies[i].collider, function () {
+						that.player.kill();
+					});
+				}
 			}
 			// Scroll background
+			this.background.ground.scrollSpeed += 1 * (realTime / 90);
 			this.background.sky.tile.tilePositionX += this.background.sky.scrollSpeed;
 			this.background.ground.tile.tilePositionX += this.background.ground.scrollSpeed;
+		} else {
+			this.enemies = [];
+			window.localStorage.setItem('score', this.score);
 		}
 	}
 }
+
+export default MainScene;
